Bfp market opt simplified (#2152)
* refactor: rm unused config and shuffle struct attrs

* refactor: pull AVG_SECONDS_PER_YEAR to const

* feat: uint128 keeperFeeBufferUsd and uint64 commitmentTime

* perf: add padding for Settlement{...}.GlobalData.maxHooksPerOrder

* refactor: constructor, comments, up storage.dump __unused{n}

* refactor: minimise block.timestamp and shuffle PerpMarket.Data

* perf: uint256->int128 for skew, funding, velocity, util

* refactor: more uint64 for commitmentTime

* refactor+perf+docs: constants and comments

* refactor: rm unused imports, libs, and explicit conversions

* refactor: rm unused custom errors

* refactor: minor rename and formatting in tests

* perf: represent {min,max}OrderAge as uint64

+++ markets/bfp-market/contracts/interfaces/IPerpAccountModule.sol
@@ -38,9 +38,9 @@ interface IPerpAccountModule {
         /// Unrealized PnL of position in USD.
         int256 pnl;
         /// Funding accrued in USD.
-        int256 accruedFunding;
+        int128 accruedFunding;
         /// Utilization accrued in USD.
-        uint256 accruedUtilization;
+        uint128 accruedUtilization;
         /// Raw Pyth entry price of position (at open or on modification).
         uint256 entryPythPrice;
         /// pd-adjusted entry price of position (at open or on modification).
